pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rl-crypto"
version = "0.1.0"
description = "PPO-based Reinforcement Learning for Cryptocurrency Futures Trading"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Ian"}
]
keywords = ["reinforcement-learning", "cryptocurrency", "trading", "ppo", "binance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "gymnasium>=0.29.0",
    "pyyaml>=6.0",
    "stable-baselines3>=2.2.0",
    "torch>=2.0.0",
    "tensorboard>=2.15.0",
    "python-binance>=1.0.19",
    "websockets>=12.0",
    "aiohttp>=3.9.0",
    "pandas-ta>=0.3.14b",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
]

[project.scripts]
rl-crypto-train = "scripts.train:main"
rl-crypto-backtest = "scripts.backtest:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "D", "UP", "B", "C4"]
ignore = ["D100", "D104"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
